requirements.txt
anyio==4.9.0
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
exceptiongroup==1.2.2
filelock==3.18.0
fsspec==2025.3.2
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pyarrow==20.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
ruff==0.11.8
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
